internal/infra/database/user_database/user_db.go
package user_database

import (
	"github.com/ViniciusSouzaDosReis/product-api/internal/entity"
	"gorm.io/gorm"
)

type User struct {
	DB *gorm.DB
}

func NewUserDB(db *gorm.DB) *User {
	return &User{DB: db}
}

func (u *User) Create(user *entity.User) error {
	return u.DB.Create(user).Error
}

func (u *User) FindByEmail(email string) (*entity.User, error) {
	var user entity.User

	// if err := u.DB.First(&user, "email = ?", email).Error; err != nil{
	// 	return nil, err
	// }
	err := u.DB.First(&user, "email = ?", email).Error
	if err != nil {
		return nil, err
	}
	return &user, nil
}
